fixed bug on 0-bubble_sort.c file

+++ 0-bubble_sort.c
@@ -6,8 +6,8 @@
  */
 void bubble_sort(int *array, size_t size)
 {
-	int i = 0;
-	int j = 1;
+	size_t i = 0;
+	size_t j = 1;
 	int temp = 0;
 
 	while(i < size)
